+++ src/main/java/nextstep/subway/line/exception/DistanceMaximumException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.subway.line.exception;
 
 public class DistanceMaximumException extends RuntimeException {
     public DistanceMaximumException() {
