+++ src/main/java/nextstep/subway/line/exception/StationDuplicateException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.subway.line.exception;
 
 public class StationDuplicateException extends RuntimeException {
     public StationDuplicateException() {
