+++ src/main/java/nextstep/subway/line/ui/LineController.java
@@ -1,6 +1,6 @@
 package nextstep.subway.line.ui;
 
-import nextstep.subway.exception.*;
+import nextstep.subway.line.exception.*;
 import nextstep.subway.line.application.LineService;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
